refactor(housekeeping): split module into smaller specialized files
* Split housekeeping module into clean, pack, indent, latexdiff components
* Improve code organization and maintainability by separating constants
* Add utility functions for common housekeeping operations
* Rename findFile to findFileInBuild for clarity
