+++ coauthor-for-vs-code/package.json
@@ -2,7 +2,7 @@
   "name": "coauthor",
   "displayName": "CoAuthor",
   "description": "CoAuthor: your never complain nor PUA nor on two-month holidays coauthor.",
-  "version": "0.13.2",
+  "version": "0.13.3",
   "publisher": "Sirui-Lu",
   "engines": {
     "vscode": "^1.93.1"
