+++ coauthor-for-vs-code/src/housekeeping/constants.ts
@@ -1,0 +1,56 @@
+export const EXCLUDED_DIRS = new Set([
+  'figs',
+  'figures',
+  'build',
+  'versions',
+  'history',
+  'notes',
+  'diffs',
+]);
+
+export const PACK_EXTENSIONS = ['.pdf', '.tex', '.txt', '.text', '.xml', '.md'];
+
+export const TEMP_EXTENSIONS = [
+  '.pdf',
+  '.aux',
+  '.bbl',
+  '.blg',
+  '.fdb_latexmk',
+  '.fls',
+  '.log',
+  '.out',
+  '.synctex.gz',
+  '.bib',
+  '.nav',
+  '.run.xml',
+  '.snm',
+  '.toc',
+  '-blx.bib',
+  'Notes.bib',
+];
+
+export const MODELS = [
+  'opus',
+  'sonnet',
+  'sonnet+',
+  'sonnet++',
+  'haiku+',
+  'haiku',
+  'o1',
+  'o1preview',
+  'o1-',
+  'gpt4ol',
+  'gpt4o',
+  'gpt4o-',
+  'gpt4t',
+  'geminiexp',
+  'gemini2p',
+  'gemini2f',
+  'gemini1p+',
+  'gemini1f+',
+  'gemini1p+OR',
+  'gemini1f+OR',
+  'llama3+OR',
+];
+
+export const HISTORY_DIR = 'History';
