+++ coauthor-for-vs-code/src/housekeeping/index.ts
@@ -1,0 +1,14 @@
+export {
+  runCleanSingle,
+  runCleanMultiple,
+  runCleanBuild,
+  runCleanOutput,
+} from './clean';
+export { runPackSingle, runPackMultiple } from './pack';
+export {
+  runPackLatexDiffVC,
+  runPackLatexDiffVCMultiple,
+  runCleanLatexDiffVC,
+  runCleanLatexDiffVCMultiple,
+} from './latexdiff';
+export { runIndentTex } from './indent';
